Include custom pixi resource class
