+++ examples/vitessce-utils.js
@@ -1,6 +1,5 @@
 import { identity } from '@flekschas/utils';
 import * as tf from '@tensorflow/tfjs-core';
-import * as PIXI from 'pixi.js';
 import { openArray } from 'zarr';
 
 tf.setBackend('cpu');
@@ -282,48 +281,3 @@ export const rgb2hsv = ([r, g, b]) => {
   }
   return [h, s, v];
 };
-
-export class CustomBufferResource extends PIXI.resources.Resource {
-  constructor(source, options) {
-    const { width, height, internalFormat, format, type } = options || {};
-
-    if (!width || !height || !internalFormat || !format || !type) {
-      throw new Error(
-        'CustomBufferResource width, height, internalFormat, format, or type invalid'
-      );
-    }
-
-    super(width, height);
-
-    this.data = source;
-    this.internalFormat = internalFormat;
-    this.format = format;
-    this.type = type;
-  }
-
-  upload(renderer, baseTexture, glTexture) {
-    const gl = renderer.gl;
-
-    gl.pixelStorei(
-      gl.UNPACK_PREMULTIPLY_ALPHA_WEBGL,
-      baseTexture.alphaMode === 1 // PIXI.ALPHA_MODES.UNPACK but `PIXI.ALPHA_MODES` are not exported
-    );
-
-    glTexture.width = baseTexture.width;
-    glTexture.height = baseTexture.height;
-
-    gl.texImage2D(
-      baseTexture.target,
-      0,
-      gl[this.internalFormat],
-      baseTexture.width,
-      baseTexture.height,
-      0,
-      gl[this.format],
-      gl[this.type],
-      this.data
-    );
-
-    return true;
-  }
-}
